add Fibonacci sequence file via upload
Enter a number (n). It shows you first n numbers of Fibonacci sequence.
In mathematics, the Fibonacci numbers are the numbers in the following integer sequence, called the Fibonacci sequence, and characterized by the fact that every number after the first two is the sum of the two preceding ones.