add file
It takes a list of words and a string. then corrects the wrong words in the string by the correct words of the list by following rules.
It does not consider two letter words. 
If a word of string differs in one letter from a word in the list, it substitutes the word of string by word of list
cas is substitute by car
cas is substitute by case
casee is substitute by case